+++ src/main/java/org/jacop/floats/search/SmallestMaxFloat.java
@@ -50,7 +50,7 @@ public class SmallestMaxFloat<T extends FloatVar> implements ComparatorVariable<
     public SmallestMaxFloat() {
     }
 
-    public int compare(float left, T var) {
+    public int compare(double left, T var) {
         double right = var.dom().max();
         if (left < right)
             return 1;
@@ -69,8 +69,8 @@ public class SmallestMaxFloat<T extends FloatVar> implements ComparatorVariable<
         return 0;
     }
 
-    public float metric(T var) {
-        return (float) var.dom().max();
+    public double metric(T var) {
+        return var.dom().max();
     }
 
 }
