+++ src/main/java/org/jacop/search/AFCMaxDeg.java
@@ -55,13 +55,13 @@ public class AFCMaxDeg<T extends Var> implements ComparatorVariable<T> {
 	this(store, store.getDecay());
     }
     
-    public AFCMaxDeg(Store store, float decay) {
+    public AFCMaxDeg(Store store, double decay) {
 	store.setAllConstraints();
 	store.afcManagement(true);
 	store.setDecay(decay);
     }
     
-    public int compare(float left, T var) {
+    public int compare(double left, T var) {
 
         double right = afcValue(var) / var.getSizeFloat();
 
@@ -95,14 +95,14 @@ public class AFCMaxDeg<T extends Var> implements ComparatorVariable<T> {
 
     }
 
-    public float metric(T var) {
+    public double metric(T var) {
 
-        return afcValue(var) / ((float)var.getSizeFloat());
+        return afcValue(var) / var.getSizeFloat();
 
     }
 
-    float afcValue(Var v) {
-	float value = 0.0f;
+    double afcValue(Var v) {
+	double value = 0.0f;
 	for (Constraint c : v.dom().constraints())
 	    value += c.afc();
 	return value;
